+++ src/main/java/vectorwing/farmersdelight/common/crafting/CookingPotRecipe.java
@@ -53,6 +53,11 @@ public class CookingPotRecipe implements Recipe<RecipeWrapper>
 	}
 
 	@Override
+	public boolean isSpecial() {
+		return true;
+	}
+
+	@Override
 	public ResourceLocation getId() {
 		return this.id;
 	}