+++ src/main/java/vectorwing/farmersdelight/common/crafting/CuttingBoardRecipe.java
@@ -48,6 +48,11 @@ public class CuttingBoardRecipe implements Recipe<RecipeWrapper>
 	}
 
 	@Override
+	public boolean isSpecial() {
+		return true;
+	}
+
+	@Override
 	public ResourceLocation getId() {
 		return this.id;
 	}