+++ changelog.md
@@ -17,6 +17,7 @@
 - Fix Chickens and Pigs not being tempted by FD crops and seeds;
 - Fix `forge:mineable/knife` being registered but not generated, causing reload problems on servers;
 - Fix some langs still referring to `_pantry` IDs instead of `_cabinet`;
+- Fix log spam regarding cooking/cutting recipe types;
 
 ## 1.0.0
 - 1.17 and 1.18 port! Codebase has been cleaned up and refactored;